Implement upload test

+++ app/src/main/java/com/example/ndtm/NdtMMeasurement.kt
@@ -34,5 +34,5 @@ fun measurementToMetrics(measurement: NdtMMeasurement?): NdtMTestMetrics? {
 }
 
 fun calcBytesPerSec(bytes: Long, usecs: Long): Double {
-    return if (usecs == 0L) 0.0 else bytes / usecs * 1e6
+    return if (usecs == 0L) 0.0 else bytes.toDouble() / usecs.toDouble() * 1e6
 }
